requirements.txt
flask
PyPDF2
pinecone
firebase_admin
python-dotenv
google-generativeai
streamlit
langchain
langchain-google-genai